API/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "firebase-admin>=6.6.0",
    "httpx>=0.28.1",
    "langchain>=0.3.15",
    "langchain-cohere>=0.3.5",
    "langchain-community>=0.3.15",
    "langchain-core>=0.3.31",
    "langchain-google-genai>=2.0.9",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tavily-python>=0.5.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.9.2",
]
